create new package 'assertions'. move TableAssert class to new package
